fix: game

+++ app/src/modules/game.py
@@ -176,6 +176,9 @@ async def handler_choice_game(chat_id, category, jawab=False):
             soal = question['soal']
             jawaban_user = jawaban_family100[chat_id]
             jawaban_soal = question['jawaban']
+            if len(jawaban_soal) == len(jawaban_family100):
+                return await handler_choice_game(chat_id, category)
+            
             format_text = f"💁 {soal}?\n"
             for index, value in enumerate(jawaban_soal):
                 user_key = list(jawaban_user.keys())[index] if index < len(jawaban_user) else None
@@ -185,7 +188,7 @@ async def handler_choice_game(chat_id, category, jawab=False):
                 else:
                     format_text += f"{index + 1}. \n"
         else:
-            return await handler_choice_game(chat_id, category, jawab=False)
+            return await handler_choice_game(chat_id, category)
         
         send_msg_jawab = await bot.send_message(chat_id, format_text, protect_content=True)
         id_msg_jwb = send_msg_jawab.id
@@ -198,7 +201,7 @@ async def handler_choice_game(chat_id, category, jawab=False):
             in_game_chat_id[chat_id]["id_msg_jwb"] = id_msg_jwb
         
         if category != "FAMILY 100":
-            return await handler_choice_game(chat_id, category, jawab=False)
+            return await handler_choice_game(chat_id, category)
         
         return 
     
@@ -503,22 +506,19 @@ async def check_answer(client, message: Message):
                 else:
                     jawaban = question["jawaban"]
                     jawab_user = message.text.strip().lower()
-                    if jawab_user in jawaban:
+                    if jawab_user in jawaban and jawab_user not in jawaban_family100[chat_id]:
                         if chat_id not in jawaban_family100:
                             jawaban_family100[chat_id] = {jawab_user: mention}
                         else:
                             jawaban_family100[chat_id].update({jawab_user: mention})
-                            
+                                               
                         await handler_choice_game(chat_id, category, jawab=True)
-                        await asyncio.sleep(2)                   
                         await bot.delete_messages(chat_id, id_msg)
-                        await bot.send_message(OWNER_ID, "pesan family 100 masuk jawaban")
-                        
-                    await bot.send_message(OWNER_ID, "pesan family 100 masuk bukan jawaban")
+                        await bot.send_message(OWNER_ID, f"{jawaban_family100}\n\npesan family 100 masuk jawaban")
                 
             if datetime.now() > end_time:
                 await client.send_message(chat_id, "<b>⏰ Waktu 3 menit telah habis!</b>", protect_content=True)
-                return await handler_choice_game(chat_id, category)        
+                return await handler_choice_game(chat_id, category)
     except:
         pass
